requirements.txt
geopy
requests
flask
sqlalchemy
sqlalchemy-utils
pymssql
pymysql
psycopg2-binary
apprise
pytz